Convert some PathErrors to DocumentNotFoundErrors
If the directory existed but was empty, a DocumentNotFoundError would be
returned (as expected). However, if the directory did not exist at all,
there would be an os.PathError for trying to open it. This error was
propagated back to the caller. Instead it should be identified, and then
replaced by a DocumentNotFoundError.

To catch this bug in the future, I put some more complete tests, but its
reproduction (at least for FileDocumentStore) is order-sensitive. If the
other tests run first, they'll create and then destroy the documents,
leaving the directory behind, which causes the empty-case to pass.