Improve watch task
I'm finally hitting my stride on this project, and pressing F5 is
finally starting to get old. This commit brings in browsersync and a
bunch of gulp.watches to make workflow smoother.

Due to cbarrick/gulp-run#17, gulp-run does not notify gulp of when the
task completes. To fix this, I reimplemented those tasks in terms of the
standard child_process module. I'm pretty pleased with how the tasks
look now.