Remove gulp-filter (for now)
Since I stopped using main-bower-files, I don't need the filtering right
now, so this dependency can be removed.

Note: in the future, I might compile min in the css task (using less),
in which case it would be useful to have the filtering again. We can
always add that dependency back when it becomes useful.