Move table class to postprocessing step
We don't want to do this as part of markdown rendering, because it makes
our XSS filter more complicated. In general, we do not want to allow the
user to specify their own class attributes. So every class attribute
generated by the renderer becomes a special case for the filter. The
fewer of those, the better.