Remove Document.Source field (YAGNI)
I really wasn't thinking when I added this field. If the Source was a
copy of the originating DocumentStore, then you would have to add a
defer doc.Source.Close() to every single Document you retrieved. That's
just awful.

Plus I kind of doubt the original use case for this field is actually
relevant. I should have YAGNId this one from the beginning. (The idea
was that, if you only had a Document, you could use the Source to pull
other Documents from the same DocumentStore. But really, when will you
have the Document and not the Store?)