Remove PathPrefix in routes
I think Gorilla's approach to subtree routing is fundamentally broken.
If you want to route to an arm of the path trie, you have to use
PathPrefix("/foo/"). This route captures "/foo/", but does not capture
"/foo", so it can't cooperate with Gorilla's StrictSlash. It also risks
entering undocumented behavior if you try to define a route for
Path("/foo") specifically. Conversely if you use PathPrefix("/foo"),
now you will also match "/food", so you have to add that to your
handler as a special case. What a mess.

Custom regular expressions, although inelegant in and of themselves,
will at least present the correct routing experience for the user. The
new routes will not match "/w" or "/w/" (which is correct since those
should redirect to the home page), and it will match both "/public" and
"/public/" (which is correct because those should both show the index of
the static files being served). Stay tuned for more refactors along this
line.

Soapbox: My preferred API for this would be r.Subtree("/public"), which
would match anything under "/public/.+". It would also be possible to
add an extra handler for r.Path("/public") to catch the root case. This
is also compatible with Gorilla's existing StrictSlash API. You can kind
of do this with subrouters, but you end up back at the same problem,
because the subrouter would need to use PathPrefix instead of the parent
router. Maybe I'm just being whiny about regexes, but if I wanted to
write custom routes all day, I wouldn't use a third-party mux in the
first place.