Add timeout to reload on goose restart
This fixes sporadic connection timeouts where a go file gets changed,
gulp rebuilds the binary and issues a browsersync reload, but the binary
is a bit slow to start. Browsersync tries to proxy the binary but it's
not listening yet, so the reload times out.

I know using sleep for this type of problem is the nuclear option, but
this is really annoying in development and I just want the bug to go
away in as little code as possible.