Refactor template partials (p3)
Now the error pages (404 and 500) use the same structure as the actual
wiki pages. That means they have a real navbar and what not.

I know there's some problematic code repetition in the HTML structure,
but Go's templates don't make it easy to factor that out. The {{yield}}
provided by unrolled/render is a hack to work around that limitation,
and it's still difficult to use. Future refactorings may find a better
way to share the <nav class="nav"><div class="container"></div></nav>
across pages, but for now it'll have to do.