Implement edit form route
The edit form route breaks into roughly three parts:

Some tweaks to main.go to actually add the route. Also, we can remove
the piece of test code that inserts a new document at startup, since it
can be done by hand once the form is implemented.

The template to serve the form. It's very rudimentary right now. The
action for the form is the POST to /w/foo/bar that was added in an
earlier commit (but I am considering moving this to another route).

The controller function to render the form. The function has to actually
fetch the document from the store, to prepopulate the content of the
form if the document does already exist. Luckily, we can reuse all the
handling logic for the existing controller function. It just happens to
fill in the exact functionality that is needed.