Move form action to /e/foo/bar with 301
I encountered a UX problem with my POST to /w/foo/bar. The user edits a
wiki page and immediately lands on that wiki page, which is good. But if
they navigate to another page and then press backspace, the browser asks
them to resend their submitted form. That's bad (it double-submits the
update AND it's also annoying).

The simplest solution is to move the POST action somewhere else and, if
the POST succeeds, perform a 301 redirect back to the wiki page where
the user can see the fruit of their labor. If they press backspace after
navigating onward, they'll go back to the wiki page without being asked
to resend the form. Much better.

What's really nice about the controller structure is that I can keep
reusing the handleDocument function for all three views. The logic is
somehow the same in every single case (the Update would be skipped on
some routes), so the only difference between the three HandlerFuncs is
the views they implement in response to the logic's outcome. I'm very
pleased with how this code looks at the moment.