Add Clear to DocumentStore API
This is part of a series of commits that will remove Revert/Truncate. I
don't really like the API guarantees these methods are trying to offer,
and I don't have a strong use case for them. However, in tests, we need
to be able to dump DocumentStores easily, so we need some way of
deleting stuff. That's what Clear is for.