Clarify Update MAX_CONTENT_SIZE
Update doesn't *have* to error out on large documents. A DocumentStore
may choose to support documents over the MAX_CONTENT_SIZE. However, the
API now notes that it must support those documents for both reads and
writes. For example, a DocumentStore is not allowed to just truncate
every document over a certain size. It should return an error, as noted,
if it can't support those large documents.