Add GetDescendants to DocumentStore API
You can read and write pages in Goose, but there's no way to discover
pages (ie to find what pages exist). Search is the main way to do this
on the modern internet, but it's also useful to have page lists.

Search is a future problem for Goose to tackle, but listing pages is a
simpler one. The first thing we need to implement this feature is a
DocumentStore method to get those pages. That's where GetDescendants
comes into the picture.